8359707: Add classfile modification code to RedefineClassHelper
